Python 3
I would like to switch the code to Python 3, but I’m not sure the best
way to make it still work well on a default Python 2.7 install (like
you would get with ArcGIS).  The only change that needs to be made is
to change the xrange calls to range, but these should stay as xrange
calls for now so that they are generators.